coding: FreeChannelの書き直しの途中
